+++ src/003.c
@@ -2,8 +2,8 @@
 #include <math.h>
 
 int main(void) {
-    unsigned long x = 600851475143;
-    unsigned long f = 0;
+    long long x = 600851475143;
+    long long f = 0;
 
     for (f = 2; f < sqrt(x); f++) {
         while (0 == x%f && x != f) {
